[MOB-76] Created LocalNotificationSync table in Room
