+++ aptoide-database/src/main/java/cm/aptoide/pt/database/room/LocalNotificationSyncDao.java
@@ -1,0 +1,18 @@
+package cm.aptoide.pt.database.room;
+
+import androidx.room.Dao;
+import androidx.room.Insert;
+import androidx.room.Query;
+import io.reactivex.Observable;
+
+import static androidx.room.OnConflictStrategy.REPLACE;
+
+@Dao public interface LocalNotificationSyncDao {
+
+  @Insert(onConflict = REPLACE) void save(RoomLocalNotificationSync roomLocalNotificationSync);
+
+  @Query("SELECT * FROM localNotificationSync WHERE notificationId = :identifier")
+  Observable<RoomLocalNotificationSync> get(String identifier);
+
+  @Query("DELETE FROM localNotificationSync WHERE notificationId = :id") void delete(String id);
+}
